Move legacy relation code into a separate lib-like file
