+++ src/constants.py
@@ -1,0 +1,11 @@
+# Copyright 2022 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""File containing constants to be used in the charm."""
+
+CLUSTER_ADMIN_USERNAME = "clusteradmin"
+SERVER_CONFIG_USERNAME = "serverconfig"
+PASSWORD_LENGTH = 24
+PEER = "database-peers"
+LEGACY_DB_ROUTER = "db-router"
+LEGACY_DB_SHARED = "shared-db"
