+++ src/mysqlsh_helpers.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
