+++ tests/integration/test_db_router.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+# Copyright 2022 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+# TODO: implement integration tests for the legacy db_router relation
