+++ tox.ini
@@ -70,6 +70,7 @@ description = Run integration tests
 deps =
     juju
     mysql-connector-python
+    protobuf==3.20.*
     pytest
     pytest-operator
     pytest-order
